some styling and context enhancements; render all data on narrow

+++ backend/graphql/PointsByGroup.ts
@@ -4,8 +4,8 @@ export const PointsByGroup = objectType({
   name: "PointsByGroup",
   definition(t) {
     t.nonNull.string("group")
-    t.nonNull.int("max_points")
-    t.nonNull.int("n_points")
+    t.nonNull.float("max_points")
+    t.nonNull.float("n_points")
     t.nonNull.float("progress")
   },
 })
@@ -14,8 +14,8 @@ export const PointsByGroupInput = inputObjectType({
   name: "PointsByGroupInput",
   definition(t) {
     t.nonNull.string("group")
-    t.nonNull.int("max_points")
-    t.nonNull.int("n_points")
+    t.nonNull.float("max_points")
+    t.nonNull.float("n_points")
     t.nonNull.float("progress")
   },
 })
